Add documentation for the message list's data source and delegate
